+++ decent_bench/library/core/cost_functions/cost_function.py
@@ -1,0 +1,90 @@
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+
+from numpy import float64
+from numpy.typing import NDArray
+
+
+class CostFunction(ABC):
+    """Used by agents to evaluate the cost and its derivatives at a certain x."""
+
+    @property
+    @abstractmethod
+    def domain_shape(self) -> tuple[int, ...]:
+        """Required shape of x."""
+
+    @property
+    @abstractmethod
+    def m_smooth(self) -> float:
+        r"""
+        Lipschitz constant of the cost function's gradient.
+
+        The gradient's Lipschitz constant m_smooth is the smallest value such that
+        :math:`\| \nabla f(\mathbf{x_1}) - \nabla f(\mathbf{x_2}) \| \leq \text{m_smooth}
+        \cdot \|\mathbf{x_1} - \mathbf{x_2}\|`
+        for all :math:`\mathbf{x_1}` and :math:`\mathbf{x_2}`.
+
+        Returns:
+            - non-negative finite number if function is L-smooth
+            - ``np.inf`` if function is differentiable everywhere but not L-smooth
+            - ``np.nan`` if function is not differentiable everywhere
+
+        """
+
+    @property
+    @abstractmethod
+    def m_cvx(self) -> float:
+        r"""
+        Convexity constant of the cost function.
+
+        The convexity constant m_cvx is the largest value such that
+        :math:`f(\mathbf{x_1}) \geq f(\mathbf{x_2})
+        + \nabla f(\mathbf{x_2})^T (\mathbf{x_1} - \mathbf{x_2})
+        + \frac{\text{m_cvx}}{2} \|\mathbf{x_1} - \mathbf{x_2}\|^2`
+        for all :math:`\mathbf{x_1}` and :math:`\mathbf{x_2}`.
+
+        Returns:
+            - positive finite number if function is strongly convex
+            - ``0`` if function is convex but not strongly convex
+            - ``np.nan`` if function is not guaranteed to be convex
+
+        """
+
+    @abstractmethod
+    def evaluate(self, x: NDArray[float64]) -> float:
+        """Evaluate function at x."""
+
+    @abstractmethod
+    def gradient(self, x: NDArray[float64]) -> NDArray[float64]:
+        """Gradient at x."""
+
+    @abstractmethod
+    def hessian(self, x: NDArray[float64]) -> NDArray[float64]:
+        """Hessian at x."""
+
+    @abstractmethod
+    def proximal(self, y: NDArray[float64], rho: float) -> NDArray[float64]:
+        r"""
+        Proximal at y.
+
+        The proximal operator is defined as:
+
+        .. include:: snippets/proximal_operator.rst
+
+        If the cost function's proximal does not have a closed form solution, it can be solved iteratively using
+        :meth:`~decent_bench.library.core.cent_algorithms.proximal_solver`.
+        """
+
+    @abstractmethod
+    def __add__(self, other: CostFunction) -> CostFunction:
+        """
+        Add another cost function to create a new one.
+
+        :class:`~.sum_cost.SumCost` can be used as the result of :meth:`__add__` by returning
+        ``SumCost([self, other])``. However, it's often more efficient to preserve the cost function type if possible.
+        For example, the addition of two :class:`~.quadratic_cost.QuadraticCost` objects benefits from returning a new
+        :class:`~.quadratic_cost.QuadraticCost` instead of a :class:`~.sum_cost.SumCost` as this preserves the closed
+        form proximal solution and only requires one evaluation instead of two when calling :meth:`evaluate`,
+        :meth:`gradient`, and :meth:`hessian`.
+        """
