+++ docs/source/decent_bench.library.core.cost_functions.rst
@@ -1,0 +1,53 @@
+decent\_bench.library.core.cost\_functions package
+==================================================
+
+Submodules
+----------
+
+decent\_bench.library.core.cost\_functions.cost\_function module
+----------------------------------------------------------------
+
+.. automodule:: decent_bench.library.core.cost_functions.cost_function
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+decent\_bench.library.core.cost\_functions.linear\_regression\_cost module
+--------------------------------------------------------------------------
+
+.. automodule:: decent_bench.library.core.cost_functions.linear_regression_cost
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+decent\_bench.library.core.cost\_functions.logistic\_regression\_cost module
+----------------------------------------------------------------------------
+
+.. automodule:: decent_bench.library.core.cost_functions.logistic_regression_cost
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+decent\_bench.library.core.cost\_functions.quadratic\_cost module
+-----------------------------------------------------------------
+
+.. automodule:: decent_bench.library.core.cost_functions.quadratic_cost
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+decent\_bench.library.core.cost\_functions.sum\_cost module
+-----------------------------------------------------------
+
+.. automodule:: decent_bench.library.core.cost_functions.sum_cost
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: decent_bench.library.core.cost_functions
+   :members:
+   :show-inheritance:
+   :undoc-members:
